Cleaning out the Solution, removing some old copies of the apps that were cluttering our branches.
