--- DonnersPlaypen-z-dev/GIDAdministration/GIDAdministration.sln
@@ -1,63 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2042
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GIDAdministration", "GIDAdministration\GIDAdministration.csproj", "{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GIDAdministration.Test", "GIDAdministration.Test\GIDAdministration.Test.csproj", "{A4AC856D-60DA-4876-8971-9D4076CD7161}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|ARM = Debug|ARM
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|ARM = Release|ARM
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|ARM.ActiveCfg = Debug|ARM
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|ARM.Build.0 = Debug|ARM
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|ARM.Deploy.0 = Debug|ARM
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|x64.ActiveCfg = Debug|x64
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|x64.Build.0 = Debug|x64
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|x64.Deploy.0 = Debug|x64
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|x86.ActiveCfg = Debug|x86
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|x86.Build.0 = Debug|x86
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Debug|x86.Deploy.0 = Debug|x86
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|ARM.ActiveCfg = Release|ARM
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|ARM.Build.0 = Release|ARM
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|ARM.Deploy.0 = Release|ARM
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|x64.ActiveCfg = Release|x64
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|x64.Build.0 = Release|x64
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|x64.Deploy.0 = Release|x64
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|x86.ActiveCfg = Release|x86
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|x86.Build.0 = Release|x86
-		{21023AF2-3FF7-40DE-87BD-AFCB0D753A48}.Release|x86.Deploy.0 = Release|x86
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|ARM.ActiveCfg = Debug|ARM
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|ARM.Build.0 = Debug|ARM
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|ARM.Deploy.0 = Debug|ARM
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|x64.ActiveCfg = Debug|x64
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|x64.Build.0 = Debug|x64
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|x64.Deploy.0 = Debug|x64
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|x86.ActiveCfg = Debug|x86
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|x86.Build.0 = Debug|x86
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Debug|x86.Deploy.0 = Debug|x86
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|ARM.ActiveCfg = Release|ARM
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|ARM.Build.0 = Release|ARM
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|ARM.Deploy.0 = Release|ARM
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|x64.ActiveCfg = Release|x64
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|x64.Build.0 = Release|x64
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|x64.Deploy.0 = Release|x64
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|x86.ActiveCfg = Release|x86
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|x86.Build.0 = Release|x86
-		{A4AC856D-60DA-4876-8971-9D4076CD7161}.Release|x86.Deploy.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {549E62B1-23F2-4986-8CBF-63333A5483B6}
-	EndGlobalSection
-EndGlobal
