--- DonnersPlaypen-z-dev/GIDAdministration/GIDAdministration/App.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="GIDAdministration.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:GIDAdministration">
-
-</Application>
