--- DonnersPlaypen-z-dev/GIDAdministration/GIDAdministration/GuildLIst.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GIDAdministration
-{
-    class GuildLIst
-    {
-        public static List<Guild> gmMain = new List<Guild>(); 
-    }
-
-    class Guild
-    {
-        public string gName = null;
-        public string gRealm = null;
-        
-    }
-}
