--- DonnersPlaypen-z-dev/GIDAdministration/GIDAdministration/MainPage.xaml
@@ -1,52 +1,0 @@
-﻿<Page
-    x:Class="GIDAdministration.MainPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:GIDAdministration"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-    <Grid Background="#FF2A1010">
-        <TextBlock x:Name="textBlock" HorizontalAlignment="Left" Text="GiD Administration Console" TextWrapping="Wrap" VerticalAlignment="Top" Height="50" Width="375" FontSize="36" RequestedTheme="Dark" FocusVisualSecondaryBrush="#9913175E" Margin="36,20,0,0">
-            <TextBlock.FocusVisualPrimaryBrush>
-                <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
-                    <GradientStop Color="Black" Offset="0"/>
-                    <GradientStop Color="#FF6E79B9" Offset="1"/>
-                </LinearGradientBrush>
-            </TextBlock.FocusVisualPrimaryBrush>
-        </TextBlock>
-        <Grid x:Name="grdNewUser" HorizontalAlignment="Left" Height="557" VerticalAlignment="Top" Width="659" Background="#FF761F1F" Margin="831,101,0,0" BorderBrush="#FF2A1010">
-            <TextBox x:Name="txtbxCharName" HorizontalAlignment="Left" Height="50" Margin="274,94,0,0" Text="Character Name: " VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-            <TextBox x:Name="txtbxCharRace" HorizontalAlignment="Left" Height="50" Margin="274,150,0,0" Text="Character Race: " VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-            <TextBox x:Name="txtbxCharClass" HorizontalAlignment="Left" Height="50" Margin="274,206,0,0" Text="Character Class: " VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-            <TextBox x:Name="txtbxCharSpec" HorizontalAlignment="Left" Height="50" Margin="274,262,0,0" Text="Character Spec: " VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-            <TextBox x:Name="txtbxRealm" HorizontalAlignment="Left" Height="50" Margin="274,318,0,0" Text="Realm: " VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-            <TextBlock x:Name="txtblkCharName" HorizontalAlignment="Left" Height="50" Margin="20,94,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark" Text="Character Name: "/>
-            <TextBlock x:Name="txtblkCharRace" HorizontalAlignment="Left" Height="50" Margin="20,150,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark"><Run Text="Character Race: "/><LineBreak/><Run/></TextBlock>
-            <TextBlock x:Name="txtBlkRealm" HorizontalAlignment="Left" Height="50" Margin="20,318,0,0" Text="Realm: " TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark"/>
-            <TextBlock x:Name="txtblkCharSpec1" HorizontalAlignment="Left" Height="50" Margin="20,262,0,0" Text="Character Spec: " TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark"/>
-            <TextBlock x:Name="txtblkCharClass" HorizontalAlignment="Left" Height="50" Margin="20,206,0,0" Text="Character Class: " TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark"/>
-            <Button x:Name="btnNewMemApply" Content="Apply" HorizontalAlignment="Left" Height="48" Margin="469,387,0,0" VerticalAlignment="Top" Width="180" FontSize="24" RequestedTheme="Dark" Background="#FF761F1F" Click="btnNewMemApply_Click" BorderBrush="#FF2A1010"/>
-            <ComboBox HorizontalAlignment="Left" Height="40" Margin="274,20,0,0" VerticalAlignment="Top" Width="348" FontSize="24">
-                <ComboBox/>
-                <ComboBoxItem Content="Ghosts in the Darkness" HorizontalAlignment="Center" VerticalAlignment="Center"/>
-                <ComboBoxItem Content="Other" HorizontalAlignment="Center" Height="100" VerticalAlignment="Center" Width="100"/>
-            </ComboBox>
-            <TextBlock x:Name="txtblkGuildName" HorizontalAlignment="Left" Height="50" Margin="20,20,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark" Text="Guild Name: "/>
-            <TextBox x:Name="txtbxCharName_Copy" HorizontalAlignment="Left" Height="50" Margin="274,94,0,0" Text="Character Name: " VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-        </Grid>
-        <Button x:Name="btnNewMember" Content="New Member" HorizontalAlignment="Left" Height="48" Margin="1310,27,0,0" VerticalAlignment="Top" Width="180" Click="btnNewMember_Click_1" FontSize="24" RequestedTheme="Dark" Background="#FF761F1F" BorderBrush="#FF2A1010"/>
-        <Button x:Name="btnAddGuild" Content="Add Guild" HorizontalAlignment="Left" Height="48" Margin="1102,27,0,0" VerticalAlignment="Top" Width="180" FontSize="24" RequestedTheme="Dark" Background="#FF761F1F" BorderBrush="#FF2A1010" Click="btnAddGuild_Click"/>
-        <Grid x:Name="grdAddGuild" HorizontalAlignment="Left" Height="557" VerticalAlignment="Top" Width="659" Background="#FF761F1F" Margin="128,101,0,0" BorderBrush="#FF2A1010">
-            <TextBlock x:Name="txtBlkAddRealm" HorizontalAlignment="Left" Height="50" Margin="20,82,0,0" Text="Realm: " TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark"/>
-            <Button x:Name="btnGuildApply" Content="Apply" HorizontalAlignment="Left" Height="48" Margin="469,387,0,0" VerticalAlignment="Top" Width="180" FontSize="24" RequestedTheme="Dark" Background="#FF761F1F" BorderBrush="#FF2A1010" Click="btnGuildApply_Click_1"/>
-            <TextBlock x:Name="txtblkAddGuildName" HorizontalAlignment="Left" Height="50" Margin="20,20,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Width="225" FontSize="24" RequestedTheme="Dark" Text="Guild Name: "/>
-            <TextBox x:Name="txtbxAddGuildName" HorizontalAlignment="Left" Height="50" Margin="274,20,0,0" Text="&lt;Insert Guild Name&gt;" VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-            <TextBox x:Name="txtbxAddRealm" HorizontalAlignment="Left" Height="50" Margin="274,82,0,0" Text="&lt;Insert Realm Name&gt;" VerticalAlignment="Top" Width="375" FontSize="24" RequestedTheme="Dark"/>
-        </Grid>
-        <TextBlock x:Name="txtblkStatus" HorizontalAlignment="Left" Height="81" Margin="659,758,0,0" Text="Status" TextWrapping="Wrap" VerticalAlignment="Top" Width="272" RequestedTheme="Dark" FontSize="48" TextAlignment="Center"/>
-
-    </Grid>
-</Page>
