+++ GIDAdministrationLegacy/GIDAdministrationLegacy/Form1.resx
