+++ GIDAdministrationLegacy/GIDAdministrationLegacy/GuildMember.cs
